% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-ICH_Mouse.R
\name{add_gene_group}
\alias{add_gene_group}
\title{add gene group}
\usage{
add_gene_group(gene_ls, group_symbol, ich_mouse)
}
\description{
add gene group
}
\section{Slots}{

\describe{
\item{\code{gene_ls}}{the gene list}

\item{\code{group_symbol}}{the gene group symbol}

\item{\code{ich_mouse}}{the class of ICH_Mouse}
}}

