% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-ICH_Mouse.R
\name{add_GO_ID_group,ANY,list-method}
\alias{add_GO_ID_group,ANY,list-method}
\title{add GO ID group}
\usage{
\S4method{add_GO_ID_group}{ANY,list}(ich_mouse, GO_id_ls)
}
\arguments{
\item{ich_mouse}{the class of ICH_Mouse}

\item{GO_id_ls}{the list of GO ID}
}
\description{
add GO ID group
}
