% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-ICH_Mouse.R
\name{annotate_the_cell_type_based_on_single_gene}
\alias{annotate_the_cell_type_based_on_single_gene}
\title{annotate the cell type based on single gene}
\usage{
annotate_the_cell_type_based_on_single_gene(ich_mouse, gene_ls)
}
\arguments{
\item{ich_mouse}{the ICH_Mouse class}

\item{gene_ls}{the gene ls used to annotate}
}
\description{
annotate the cell type based on single gene
}
