% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-Graph_Database.R
\name{combine_plots,ANY,Graph_Database-method}
\alias{combine_plots,ANY,Graph_Database-method}
\title{combine plots}
\usage{
\S4method{combine_plots}{ANY,Graph_Database}(
  plot_name_ls,
  graph_database = ICHMousewch::Graph_Database(),
  combination_num = 2,
  graph_name = character()
)
}
\arguments{
\item{plot_name_ls}{the plot for combination}

\item{graph_database}{the class of graph database}

\item{combination_num}{the number to combination}

\item{graph_name}{the graph name}
}
\description{
combine plots
}
