% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-Graph_Database.R
\name{combine_plots,Graph_Database-method}
\alias{combine_plots,Graph_Database-method}
\title{combine plots}
\usage{
\S4method{combine_plots}{Graph_Database}(graph_database, plot_name_ls, combination_num = 2)
}
\arguments{
\item{graph_database}{the class of graph database}

\item{plot_name_ls}{the plot for combination}

\item{combination_num}{the number to combination}
}
\description{
combine plots
}
