% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-general.R
\name{combined_matrix_on_column}
\alias{combined_matrix_on_column}
\title{combined matrix on column}
\usage{
combined_matrix_on_column(matrix_ls)
}
\arguments{
\item{matrix_ls}{the list of matrix}
}
\description{
combined matrix on column
}
