% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-ICH_Mouse.R
\name{conduct_GO_enrichment}
\alias{conduct_GO_enrichment}
\title{conduct GO enrichment and cluster GO term}
\usage{
conduct_GO_enrichment(ich_mouse, gene_set_name = "edge-normal")
}
\arguments{
\item{ich_mouse}{the class of ICH_Mouse}

\item{gene_set_name}{the gene set to find expression genes}
}
\description{
conduct GO enrichment and cluster GO term
}
