% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-Graph_Database.R
\name{create_combined_plots_set,ICH_Mouse-method}
\alias{create_combined_plots_set,ICH_Mouse-method}
\title{create combined plots set}
\usage{
\S4method{create_combined_plots_set}{ICH_Mouse}(ich_mouse)
}
\arguments{
\item{ich_mouse}{the class of ICH_Mouse}
}
\description{
create combined plots set
}
