% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-Graph_Database.R
\name{create_combined_plots_set}
\alias{create_combined_plots_set}
\title{create combined plots set}
\usage{
create_combined_plots_set(ich_mouse)
}
\arguments{
\item{ich_mouse}{the class of ICH_Mouse}
}
\description{
create combined plots set
}
