% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-Enrichment_Set.R
\name{Create_Enrichment_Set}
\alias{Create_Enrichment_Set}
\title{constructor of Enrichment_Set}
\usage{
Create_Enrichment_Set(ich_mouse, initialization = TRUE)
}
\arguments{
\item{ich_mouse}{the ICH_Mouse class}
}
\description{
constructor of Enrichment_Set
}
