% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-Gene_Spatial_Pattern.R
\name{Create_Gene_Spatial_Pattern}
\alias{Create_Gene_Spatial_Pattern}
\title{create class of Gene_Spatial_Pattern}
\usage{
Create_Gene_Spatial_Pattern(ich_mouse, gene_ls, pattern_na, bin_scale = 5000)
}
\arguments{
\item{ich_mouse}{the class of ich_mouse}

\item{gene_ls}{the genes for classfication}

\item{pattern_na}{the spatial pattern name}

\item{bin_scale}{the scale of the bin}
}
\description{
create class of Gene_Spatial_Pattern
}
