% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-Spatial_Image.R
\name{create_progeny_score_map,ANY-method}
\alias{create_progeny_score_map,ANY-method}
\title{create progeny score map}
\usage{
\S4method{create_progeny_score_map}{ANY}(image_set_name)
}
\arguments{
\item{image_set_name}{the name of the image set}
}
\description{
create progeny score map
}
