% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-Spatial_Image.R
\name{create_single_gene_contrasting_spatial_image,ICH_Mouse,ICH_Mouse,character,character-method}
\alias{create_single_gene_contrasting_spatial_image,ICH_Mouse,ICH_Mouse,character,character-method}
\title{create single contrasting gene spatial image}
\usage{
\S4method{create_single_gene_contrasting_spatial_image}{ICH_Mouse,ICH_Mouse,character,character}(
  ich_mouse1,
  ich_mouse2,
  gene_ls,
  image_set_name,
  show_background_image = TRUE
)
}
\arguments{
\item{ich_mouse1}{the class of ICH_Mouse}

\item{ich_mouse2}{the class of ICH_Mouse}

\item{gene_ls}{a gene list}

\item{image_set_name}{the name of the image set}

\item{show_background_image}{whether to show background image}
}
\description{
create single contrasting gene spatial image
}
