% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-Spatial_Image.R
\name{create_single_gene_contrasting_spatial_image}
\alias{create_single_gene_contrasting_spatial_image}
\title{create single gene contrasting spatial image}
\usage{
create_single_gene_contrasting_spatial_image(
  ich_mouse1,
  ich_mouse2,
  gene_ls,
  image_set_name,
  show_background_image = TRUE
)
}
\arguments{
\item{ich_mouse1}{the class of ICH_Mouse}

\item{ich_mouse2}{the class of ICH_Mouse}

\item{gene_ls}{a gene list}

\item{image_set_name}{the name of the image set}

\item{show_background_image}{whether to show background image}
}
\description{
create single gene contrasting spatial image
}
