% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-Spatial_Image.R
\name{create_single_gene_distribution_map}
\alias{create_single_gene_distribution_map}
\title{create single gene distribution map}
\usage{
create_single_gene_distribution_map(ich_mouse, gene_ls, image_set_name)
}
\arguments{
\item{ich_mouse}{the class of ICH_Mouse}

\item{gene_ls}{a gene list}

\item{image_set_name}{the name of the image set}
}
\description{
create single gene distribution map
}
