% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-Spatial_Image.R
\name{create_single_gene_reference_mouse_cell_barchart,character,character-method}
\alias{create_single_gene_reference_mouse_cell_barchart,character,character-method}
\title{create single gene reference mouse cell barchart}
\usage{
\S4method{create_single_gene_reference_mouse_cell_barchart}{character,character}(gene_ls, image_set_name)
}
\arguments{
\item{gene_ls}{a gene list}

\item{image_set_name}{the name of the image set}
}
\description{
create single gene reference mouse cell barchart
}
