% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-Spatial_Image.R
\name{create_single_gene_reference_mouse_immume_cell_barchart}
\alias{create_single_gene_reference_mouse_immume_cell_barchart}
\title{create single gene reference mouse immune cell barchart}
\usage{
create_single_gene_reference_mouse_immume_cell_barchart(
  gene_ls,
  image_set_name
)
}
\arguments{
\item{gene_ls}{a gene list}

\item{image_set_name}{the name of the image set}
}
\description{
create single gene reference mouse immune cell barchart
}
