% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-Spatial_Image.R
\name{create_single_gene_spatial_image,ICH_Mouse,character,character-method}
\alias{create_single_gene_spatial_image,ICH_Mouse,character,character-method}
\title{create single gene spatial image}
\usage{
\S4method{create_single_gene_spatial_image}{ICH_Mouse,character,character}(
  ich_mouse,
  gene_ls,
  image_set_name,
  show_background_image = TRUE,
  gradient = FALSE
)
}
\arguments{
\item{gene_ls}{a gene list}

\item{image_set_name}{the name of the image set}

\item{show_background_image}{whether to show background image}

\item{ICH_Mouse}{the class of ICH_Mouse}
}
\description{
create single gene spatial image
}
