% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-ICH_Mouse.R
\name{create_single_gene_spatial_image,ICH_Mouse,character,logical-method}
\alias{create_single_gene_spatial_image,ICH_Mouse,character,logical-method}
\title{create single gene spatial image}
\usage{
\S4method{create_single_gene_spatial_image}{ICH_Mouse,character,logical}(ich_mouse, gene_ls, show_background_image)
}
\arguments{
\item{gene_ls}{the gene list to create spatial image}

\item{ICH_Mouse}{the class of ICH_Mouse}
}
\description{
create single gene spatial image
}
