% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-Spatial_Image.R
\name{create_single_gene_spatial_image}
\alias{create_single_gene_spatial_image}
\title{create single gene spatial image}
\usage{
create_single_gene_spatial_image(
  ich_mouse,
  gene_ls,
  image_set_name,
  show_background_image = TRUE,
  gradient = FALSE
)
}
\arguments{
\item{ich_mouse}{the class of ICH_Mouse}

\item{gene_ls}{a gene list}

\item{image_set_name}{the name of the image set}

\item{show_background_image}{whether to show background image}

\item{gradient}{whether to use gradient color}
}
\description{
create single gene spatial image
}
