% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-Spatial_Chunk.R
\name{Create_Spatial_Chunk}
\alias{Create_Spatial_Chunk}
\title{initialize the class of Spatial_Chunk}
\usage{
Create_Spatial_Chunk(chunk_set_name, barcodes, bin_scale = 5000)
}
\arguments{
\item{chunk_set_name}{the name of the chunk set}

\item{barcodes}{the barcodes for chunking}

\item{bin_scale}{the scale of the chunk}
}
\description{
initialize the class of Spatial_Chunk
}
