% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-ICH_Mouse.R
\name{create_spatial_image_with_highlighted_clusters,Hematoma,character,numeric-method}
\alias{create_spatial_image_with_highlighted_clusters,Hematoma,character,numeric-method}
\title{create spatial image with highlighted clusters}
\usage{
\S4method{create_spatial_image_with_highlighted_clusters}{Hematoma,character,numeric}(
  hematoma,
  cluster_symbol,
  cluster_ls
)
}
\arguments{
\item{hematoma}{the Hematoma Class}

\item{cluster_symbol}{the cluster symbol for plotiing}

\item{cluster_ls}{the cluster list for plot}
}
\description{
create spatial image with highlighted clusters
}
