% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-Hematoma.R
\name{create_spatial_image_with_highlighted_clusters}
\alias{create_spatial_image_with_highlighted_clusters}
\title{generic function for create spatial image with highlighted clusters}
\usage{
create_spatial_image_with_highlighted_clusters(
  hematoma,
  cluster_symbol,
  cluster_ls
)
}
\arguments{
\item{hematoma}{the class of HematomaICHMousewch}

\item{cluster_symbol}{the cluster symbol for plotiing}

\item{cluster_ls}{the cluster list for plot}
}
\description{
generic function for create spatial image with highlighted clusters
}
