% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-enrichment_analysis.R
\name{.add_go_term_condition_information}
\alias{.add_go_term_condition_information}
\title{add go term condition information}
\usage{
.add_go_term_condition_information(gene_info_df, go_term_list, go_result)
}
\arguments{
\item{gene_info_df}{the gene information dataframe}

\item{go_term_list}{the GO term list}

\item{go_result}{the GO analysis result}
}
\description{
add go term condition information
}
