% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-enrichment_analysis.R
\name{.adjust_iteration_cluster_results}
\alias{.adjust_iteration_cluster_results}
\title{adjust iteration cluster results}
\usage{
.adjust_iteration_cluster_results(GO_results)
}
\arguments{
\item{GO_results}{the results of GO enrichment}
}
\description{
adjust iteration cluster results
}
