% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-general.R
\name{.adjust_special_block_GO_group_order}
\alias{.adjust_special_block_GO_group_order}
\title{adjust special block GO group order}
\usage{
.adjust_special_block_GO_group_order(GO_cluster, GO_group)
}
\arguments{
\item{GO_cluster}{the GO_cluster}

\item{GO_group}{the GO group}
}
\description{
adjust special block GO group order
}
