% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-spatial_pattern.R
\name{.calculate_chunk_characteristic_value}
\alias{.calculate_chunk_characteristic_value}
\title{calculate chunk characteristic value}
\usage{
.calculate_chunk_characteristic_value(
  seu_meta,
  raw_count_matrix,
  spatial_chunks,
  gene_ls,
  kernel_function = "Gaussian"
)
}
\arguments{
\item{seu_meta}{the Seurat Object metadata}

\item{raw_count_matrix}{the raw count matrix}

\item{spatial_chunks}{the chunks for calculation}

\item{gene_ls}{the gene list from expression symbol matrix}

\item{kernel_function}{choose the kernel function}
}
\description{
calculate chunk characteristic value
}
