% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-general.R
\name{.calculate_pathway_features}
\alias{.calculate_pathway_features}
\title{calculate pathway features}
\usage{
.calculate_pathway_features(pathway_ID, ich_mouse)
}
\arguments{
\item{pathway_ID}{the wikipathway ID}

\item{ich_mouse}{the class of ICH_Mouse}
}
\description{
calculate pathway features
}
