% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-pathway_analysis.R
\name{.calculate_progeny_score}
\alias{.calculate_progeny_score}
\title{calculate progeny score}
\usage{
.calculate_progeny_score(ich_mouse)
}
\arguments{
\item{ich_mouse}{the class of ICH_Mouse}
}
\description{
calculate progeny score
}
