% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-enrichment_analysis.R
\name{.calculate_single_gene_average_expression}
\alias{.calculate_single_gene_average_expression}
\title{calculate single gene average expression}
\usage{
.calculate_single_gene_average_expression(ich_mosue, gene_ls)
}
\arguments{
\item{ich_mosue}{the class of ICH_Mouse}

\item{gene_ls}{the gene list}
}
\description{
calculate single gene average expression
}
