% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-general.R
\name{.choose_bins_for_histogram}
\alias{.choose_bins_for_histogram}
\title{choose bins for histogram}
\usage{
.choose_bins_for_histogram(his_dataset, reso = 0.05, bins = 0)
}
\arguments{
\item{his_dataset}{histogram dataset}
}
\description{
choose bins for histogram
}
