% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-identify_hematoma.R
\name{.choose_center_edge_symbol}
\alias{.choose_center_edge_symbol}
\title{choose center edge symbol}
\usage{
.choose_center_edge_symbol(hematoma)
}
\arguments{
\item{hematoma}{the class of Hematoma}
}
\description{
choose center edge symbol
}
