% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-identify_hematoma.R
\name{.choose_hematoma_symbol}
\alias{.choose_hematoma_symbol}
\title{choose hematoma symbol}
\usage{
.choose_hematoma_symbol(hematoma)
}
\arguments{
\item{hematoma}{the class of Hematoma}
}
\description{
choose hematoma symbol
}
