% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-enrichment_analysis.R
\name{.cluster_GO_terms}
\alias{.cluster_GO_terms}
\title{cluster the GO term by dynamicTreeCut}
\usage{
.cluster_GO_terms(GO_results)
}
\arguments{
\item{GO_results}{the results of GO enrichment}
}
\description{
cluster the GO term by dynamicTreeCut
}
