% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-graph_spatial_image.R
\name{.combine_gene_distribution_map}
\alias{.combine_gene_distribution_map}
\title{combine gene distribution map}
\usage{
.combine_gene_distribution_map(gene_set, ...)
}
\arguments{
\item{gene_set}{the gene set}
}
\description{
combine gene distribution map
}
