% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-enrichment_analysis.R
\name{.combine_GO_cluster_based_on_cluster_annotation}
\alias{.combine_GO_cluster_based_on_cluster_annotation}
\title{combine GO cluster based on cluster annotation}
\usage{
.combine_GO_cluster_based_on_cluster_annotation(GO_results, cluster_annotation)
}
\arguments{
\item{GO_results}{GO enrichment results}

\item{cluster_annotation}{the cluster annotation}
}
\description{
combine GO cluster based on cluster annotation
}
