% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-enrichment_analysis.R
\name{.combine_GO_enrichment_heatmap}
\alias{.combine_GO_enrichment_heatmap}
\title{combine a GO enrichment heatmap}
\usage{
.combine_GO_enrichment_heatmap(GO_results_ls)
}
\arguments{
\item{GO_results_ls}{the GO results list}
}
\description{
combine a GO enrichment heatmap
}
