% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-combined_plots.R
\name{.combine_spatial_image_and_umap}
\alias{.combine_spatial_image_and_umap}
\title{combine spatial image and umap}
\usage{
.combine_spatial_image_and_umap(ich_mouse)
}
\arguments{
\item{ich_mouse}{the ICH_Mouse}
}
\description{
combine spatial image and umap
}
