% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-combined_plots.R
\name{.combined_GMM_barchat_and_count_distribution_map}
\alias{.combined_GMM_barchat_and_count_distribution_map}
\title{combined GMM barchart and count distribution map}
\usage{
.combined_GMM_barchat_and_count_distribution_map(ich_mouse)
}
\arguments{
\item{ich_mouse}{the ICH_Mouse}
}
\description{
combined GMM barchart and count distribution map
}
