% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-combined_plots.R
\name{.combined_GO_overview_plot}
\alias{.combined_GO_overview_plot}
\title{combined GO overview plot}
\usage{
.combined_GO_overview_plot(ich_mouse, group_symbol)
}
\arguments{
\item{ich_mouse}{the ICH_Mouse}

\item{group_symbol}{the group symbol}
}
\description{
combined GO overview plot
}
