% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-combined_plots.R
\name{.combined_volcano_plot_and_venn_diagram}
\alias{.combined_volcano_plot_and_venn_diagram}
\title{combined volcano plot and venn diagram}
\usage{
.combined_volcano_plot_and_venn_diagram(ich_mouse)
}
\arguments{
\item{ich_mouse}{the ICH_Mouse}
}
\description{
combined volcano plot and venn diagram
}
