% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-cluster.R
\name{.conduct_bayesspace_cluster}
\alias{.conduct_bayesspace_cluster}
\title{conduct bayesspace cluster}
\usage{
.conduct_bayesspace_cluster(
  seu_meta,
  raw_count_matrix,
  symbol_genes,
  adjust_symbol = "center"
)
}
\arguments{
\item{seu_meta}{the Seurat metadata}

\item{raw_count_matrix}{the raw_count_matrix}

\item{symbol_genes}{the symbol genes for cluster}

\item{adjust_symbol}{the adjust symbol}
}
\description{
conduct bayesspace cluster
}
