% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-cluster.R
\name{.conduct_GMM_classification}
\alias{.conduct_GMM_classification}
\title{conduct GMM classification}
\usage{
.conduct_GMM_classification(original_seu_metadata, in_tissue_barcode)
}
\arguments{
\item{original_seu_metadata}{original Seurat Object metadata}

\item{in_tissue_barcode}{the barcodes in tissue}
}
\description{
conduct GMM classification
}
