% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-GO_analysis.R
\name{.conduct_GO_enrichment}
\alias{.conduct_GO_enrichment}
\title{conduct GO enrichment}
\usage{
.conduct_GO_enrichment(gene_ls)
}
\arguments{
\item{gene_ls}{a gene list}
}
\description{
conduct GO enrichment
}
