% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-cluster.R
\name{.conduct_HMRF_cluster}
\alias{.conduct_HMRF_cluster}
\title{conduct HMRF cluster}
\usage{
.conduct_HMRF_cluster(
  giotto_instruction,
  seu_meta,
  raw_count_matrix,
  symbol_genes
)
}
\arguments{
\item{giotto_instruction}{the instruction of Giotto Object}

\item{seu_meta}{the Seurat Object metadata}

\item{raw_count_matrix}{the raw count matrix}

\item{symbol_genes}{the symbol genes}
}
\description{
conduct HMRF cluster
}
