% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-enrichment_analysis.R
\name{.conduct_iteration_cluster_on_GO_terms}
\alias{.conduct_iteration_cluster_on_GO_terms}
\title{conduct iteration cluster on GO terms}
\usage{
.conduct_iteration_cluster_on_GO_terms(GO_results)
}
\arguments{
\item{GO_results}{the results of GO enrichment}
}
\description{
conduct iteration cluster on GO terms
}
