% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-GO_analysis.R
\name{.conduct_KEEG_enrichment}
\alias{.conduct_KEEG_enrichment}
\title{conduct KEEG enrichment}
\usage{
.conduct_KEEG_enrichment(gene_ls)
}
\arguments{
\item{gene_ls}{a gene list}
}
\description{
conduct KEEG enrichment
}
