% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-enrichment_analysis.R
\name{.conduct_KEGG_enrichment}
\alias{.conduct_KEGG_enrichment}
\title{conduct KEGG enrichment}
\usage{
.conduct_KEGG_enrichment(gene_ls, filtered_genes)
}
\arguments{
\item{gene_ls}{a gene list}

\item{filtered_genes}{the filtered genes}
}
\description{
conduct KEGG enrichment
}
