% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-cluster.R
\name{.conduct_Louvain_cluster_based_on_tissue_position}
\alias{.conduct_Louvain_cluster_based_on_tissue_position}
\title{conduct Louvain cluster based on tissue position matrix}
\usage{
.conduct_Louvain_cluster_based_on_tissue_position(
  original_seu_metadata,
  raw_count_matrix
)
}
\arguments{
\item{original_seu_metadata}{original Seurat Object metadata}

\item{raw_count_matrix}{the matrix of raw count dataset}
}
\description{
conduct Louvain cluster based on tissue position matrix
}
