% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-cluster.R
\name{.conduct_Louvain_cluster_on_custom_features}
\alias{.conduct_Louvain_cluster_on_custom_features}
\title{conduct Louvain cluster on custom features}
\usage{
.conduct_Louvain_cluster_on_custom_features(
  seu_meta,
  raw_count_matrix,
  custom_features
)
}
\arguments{
\item{seu_meta}{the Seurat Object metadata}

\item{raw_count_matrix}{the row count matrix}

\item{custom_features}{the custom features}
}
\description{
conduct Louvain cluster on custom features
}
