% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-general.R
\name{.conduct_statistic_test_on_rate}
\alias{.conduct_statistic_test_on_rate}
\title{conduct statistic test on rate}
\usage{
.conduct_statistic_test_on_rate(sample_dt)
}
\arguments{
\item{sample_dt}{the sample for analysis}
}
\description{
conduct statistic test on rate
}
