% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-general.R
\name{.construct_GO_similarity_heatmap_dataset}
\alias{.construct_GO_similarity_heatmap_dataset}
\title{construct GO similarity heatmap dataset}
\usage{
.construct_GO_similarity_heatmap_dataset(GO_results)
}
\arguments{
\item{GO_results}{the GO enrichment result}
}
\description{
construct GO similarity heatmap dataset
}
