% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-graph.R
\name{.create_categorical_network_plot}
\alias{.create_categorical_network_plot}
\title{create Categorical Network Plot of GO term}
\usage{
.create_categorical_network_plot(GO_results)
}
\arguments{
\item{GO_results}{the results of GO enrichment}
}
\description{
create Categorical Network Plot of GO term
}
