% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-spatial_pattern.R
\name{.create_chunks}
\alias{.create_chunks}
\title{create chunks}
\usage{
.create_chunks(total_num, bin_num)
}
\arguments{
\item{total_num}{the total number for creating chunks}

\item{bin_num}{the number of chunks}
}
\description{
create chunks
}
