% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-graph.R
\name{.create_gene_group_barchart}
\alias{.create_gene_group_barchart}
\title{create gene group barchart}
\usage{
.create_gene_group_barchart(ich_mouse, gene_name)
}
\arguments{
\item{ich_mouse}{the class of ICH_Mouse}

\item{gene_name}{the gene name}
}
\description{
create gene group barchart
}
