% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-graph.R
\name{.create_giotto_instruction}
\alias{.create_giotto_instruction}
\title{create Giotto Instruction}
\usage{
.create_giotto_instruction(python_path, results_folder)
}
\arguments{
\item{python_path}{the path to a python which the Giotto can use}

\item{results_folder}{the folder for Giotto save plots}
}
\description{
create Giotto Instruction
}
