% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-graph.R
\name{.create_GO_cluster_graph_combination_set}
\alias{.create_GO_cluster_graph_combination_set}
\title{create GO cluster graph combination set}
\usage{
.create_GO_cluster_graph_combination_set(
  GO_ID_group,
  GO_enrichment,
  seu_meta,
  raw_count_matrix,
  filtered_genes,
  filtered_barcodes,
  diff_expr_gene
)
}
\arguments{
\item{GO_ID_group}{the GO ID group}

\item{GO_enrichment}{the result of GO enrichment}

\item{seu_meta}{the Seurat metadata}

\item{raw_count_matrix}{the raw count matrix}

\item{filtered_genes}{the filtered genes}

\item{filtered_barcodes}{the filtered barcodes}

\item{diff_expr_gene}{the differential expression genes}
}
\description{
create GO cluster graph combination set
}
