% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-graph.R
\name{.create_GO_result_custom_barchart}
\alias{.create_GO_result_custom_barchart}
\title{create GO result custom barchart}
\usage{
.create_GO_result_custom_barchart(ich_mouse, GO_term_ls)
}
\arguments{
\item{ich_mouse}{the class of ICH_Mouse}

\item{GO_term_ls}{the list of GO term}
}
\description{
create GO result custom barchart
}
