% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-graph.R
\name{.create_GO_result_overview_barchart}
\alias{.create_GO_result_overview_barchart}
\title{create GO result overview barchart}
\usage{
.create_GO_result_overview_barchart(ich_mouse)
}
\arguments{
\item{ich_mouse}{the class of ICH_Mouse}
}
\description{
create GO result overview barchart
}
