% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-graph.R
\name{.create_heatmap_of_mouse_cell_reference_dataset}
\alias{.create_heatmap_of_mouse_cell_reference_dataset}
\title{create heatmap of mouse cell reference dataset}
\usage{
.create_heatmap_of_mouse_cell_reference_dataset(gene_ls, reference_data_symbol)
}
\arguments{
\item{gene_ls}{the gene list for heatmap}

\item{reference_data_symbol}{the reference data symbol}
}
\description{
create heatmap of mouse cell reference dataset
}
