% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-graph.R
\name{.create_ncount_nfeature_histogram}
\alias{.create_ncount_nfeature_histogram}
\title{create ncount and nfeature histogram}
\usage{
.create_ncount_nfeature_histogram(seu_metadata, plotting_title, reso = 0.05)
}
\arguments{
\item{seu_metadata}{the Seurat Object metadata}

\item{plotting_title}{the title of plotting}
}
\description{
create ncount and nfeature histogram
}
