% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-pathway_analysis.R
\name{.create_progeny_pathway_score_map}
\alias{.create_progeny_pathway_score_map}
\title{create progeny pathway score map}
\usage{
.create_progeny_pathway_score_map(progeny_score_df)
}
\arguments{
\item{progeny_score_df}{the progeny score dataframe}
}
\description{
create progeny pathway score map
}
