% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-reference_dataset.R
\name{.create_public_reference_plotting_dataset}
\alias{.create_public_reference_plotting_dataset}
\title{create public plotting dataset}
\usage{
.create_public_reference_plotting_dataset()
}
\description{
create public plotting dataset
}
