% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-reference_dataset.R
\name{.create_public_reference_plotting_dataset_mouse_cell}
\alias{.create_public_reference_plotting_dataset_mouse_cell}
\title{create public plotting dataset}
\usage{
.create_public_reference_plotting_dataset_mouse_cell()
}
\description{
create public plotting dataset
}
