% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-reference_dataset.R
\name{.create_public_reference_plotting_dataset_mouse_immune_cell}
\alias{.create_public_reference_plotting_dataset_mouse_immune_cell}
\title{create public plotting dataset mouse immune cell}
\usage{
.create_public_reference_plotting_dataset_mouse_immune_cell()
}
\description{
create public plotting dataset mouse immune cell
}
