% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-graph.R
\name{.create_single_gene_infotmation_heatmap}
\alias{.create_single_gene_infotmation_heatmap}
\title{create single gene information heatmap}
\usage{
.create_single_gene_infotmation_heatmap(
  ich_mouse,
  gene_ls,
  go_term_ls,
  GO_symbol = "filtered_total-diff_expr_genes"
)
}
\arguments{
\item{ich_mouse}{the class of ICH_Mouse}

\item{gene_ls}{the gene list}

\item{go_term_ls}{the GO term list}

\item{GO_symbol}{the GO analysis symbol}
}
\description{
create single gene information heatmap
}
