% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-Spatial_Image.R
\name{.Create_Spatial_Image}
\alias{.Create_Spatial_Image}
\title{Spatial Image constructor}
\usage{
.Create_Spatial_Image(image_set_name, spatial_image_ls, initialization = FALSE)
}
\arguments{
\item{image_set_name}{the name of image set}

\item{spatial_image_ls}{a list contain a set of spatial image}
}
\description{
Spatial Image constructor
}
