% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-graph.R
\name{.create_umap_plot}
\alias{.create_umap_plot}
\title{create umap plot}
\usage{
.create_umap_plot(
  raw_count_matrix,
  seu_meta,
  diff_expr_genes,
  filter_genes,
  umap_feature = "diff_expr_genes"
)
}
\arguments{
\item{raw_count_matrix}{the raw count matrix}

\item{seu_meta}{the seurat metadata}

\item{diff_expr_genes}{the different expression genes}

\item{filter_genes}{the filter genes}

\item{umap_feature}{the features for umap}
}
\description{
create umap plot
}
