% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-graph.R
\name{.create_Venn_Diagram}
\alias{.create_Venn_Diagram}
\title{create Venn Diagram}
\usage{
.create_Venn_Diagram(ich_mouse)
}
\arguments{
\item{ich_mouse}{the class of ICH_Mouse}
}
\description{
create Venn Diagram
}
