% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-graph.R
\name{.create_violin_plot_of_quality_control}
\alias{.create_violin_plot_of_quality_control}
\title{create violin plot of quality control}
\usage{
.create_violin_plot_of_quality_control(
  seu_metadata,
  filtered_barcodes,
  plotting_var
)
}
\arguments{
\item{seu_metadata}{the Seurat Object metadata}

\item{filtered_barcodes}{the filtered barcodes}

\item{plotting_var}{the variable for plotting}
}
\description{
create violin plot of quality control
}
