% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-reference_dataset.R
\name{.diff_expr_genes_cell_type_one_to_other}
\alias{.diff_expr_genes_cell_type_one_to_other}
\title{differential exoression genes one type to other type according to mouse RNA seq}
\usage{
.diff_expr_genes_cell_type_one_to_other()
}
\description{
differential exoression genes one type to other type according to mouse RNA seq
}
