% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-reference_dataset.R
\name{.diff_expr_genes_one_type_to_one_type}
\alias{.diff_expr_genes_one_type_to_one_type}
\title{differential exoression genes one type to one type}
\usage{
.diff_expr_genes_one_type_to_one_type()
}
\description{
differential exoression genes one type to one type
}
