% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-reference_dataset.R
\name{.download_gene_id_information}
\alias{.download_gene_id_information}
\title{download gene id information}
\usage{
.download_gene_id_information()
}
\description{
download gene id information
}
