% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-general.R
\name{.extract_group_GO_result}
\alias{.extract_group_GO_result}
\title{extract_group_GO_result}
\usage{
.extract_group_GO_result(GO_result, gene_set)
}
\arguments{
\item{GO_result}{the GO enrichment result}

\item{gene_set}{the gene list}
}
\description{
extract_group_GO_result
}
