% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-general.R
\name{.filter_diff_expr_genes}
\alias{.filter_diff_expr_genes}
\title{filter differential expression genes}
\usage{
.filter_diff_expr_genes(diff_expr_genes)
}
\arguments{
\item{diff_expr_genes}{the differential expression genes}
}
\description{
filter differential expression genes
}
