% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-general.R
\name{.filter_GO_term_based_on_contained_matrix}
\alias{.filter_GO_term_based_on_contained_matrix}
\title{filter GO term based on contained matrix}
\usage{
.filter_GO_term_based_on_contained_matrix(GO_ID_group, GO_enrichment)
}
\arguments{
\item{GO_ID_group}{the GO terms}

\item{GO_enrichment}{the result of GO enrichment}
}
\description{
filter GO term based on contained matrix
}
