% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-general.R
\name{.find_filtered_genes}
\alias{.find_filtered_genes}
\title{find filtered genes}
\usage{
.find_filtered_genes(raw_count_matrix, in_tissue_barcodes)
}
\arguments{
\item{raw_count_matrix}{the matrix of raw count dataset}

\item{in_tissue_barcodes}{the barcodes in tissue}
}
\description{
find filtered genes
}
