% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-general.R
\name{.find_single_gene_go_terms}
\alias{.find_single_gene_go_terms}
\title{find single gene go terms}
\usage{
.find_single_gene_go_terms(
  gene_ls,
  ich_mouse,
  go_result_symbol = "filtered_total-diff_expr_genes"
)
}
\arguments{
\item{gene_ls}{the gene list}

\item{ich_mouse}{the class of ICH_Mouse}

\item{go_result_symbol}{the GO result symbol}
}
\description{
find single gene go terms
}
