% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-ANOVA.R
\name{.find_symbol_genes}
\alias{.find_symbol_genes}
\title{find symbol genes according to cluster}
\usage{
.find_symbol_genes(
  raw_count_matrix,
  seu_metadata_with_cluster_symbol,
  filtered_genes,
  cluster_symbol
)
}
\arguments{
\item{raw_count_matrix}{the matrix of raw count dataset}

\item{seu_metadata_with_cluster_symbol}{the Seurat Object metadata with cluster symbol}

\item{filtered_genes}{the gene that been filtered}

\item{cluster_symbol}{the symbol of cluster}
}
\description{
find symbol genes according to cluster
}
