% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-spatial_pattern.R
\name{.generate_expression_symbol_matrix}
\alias{.generate_expression_symbol_matrix}
\title{generate expression symbol matrix}
\usage{
.generate_expression_symbol_matrix(barcodes, raw_count_matrix, gene_ls)
}
\arguments{
\item{barcodes}{the barcodes of a chunk}

\item{raw_count_matrix}{the raw count matrix}

\item{gene_ls}{the gene list from expression symbol matrix}
}
\description{
generate expression symbol matrix
}
