% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-general.R
\name{.generate_gene_GO_group}
\alias{.generate_gene_GO_group}
\title{generate gene GO group}
\usage{
.generate_gene_GO_group(ich_mouse, gene_ls)
}
\arguments{
\item{ich_mouse}{the class of ICH_Mouse}

\item{gene_ls}{the gene list}
}
\description{
generate gene GO group
}
