% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-general.R
\name{.generate_GO_term_gene_contained_matrix}
\alias{.generate_GO_term_gene_contained_matrix}
\title{generate GO term gene contained matrix}
\usage{
.generate_GO_term_gene_contained_matrix(GO_ID_group, GO_enrichment)
}
\arguments{
\item{GO_ID_group}{the GO terms}

\item{GO_enrichment}{the result of GO enrichment}
}
\description{
generate GO term gene contained matrix
}
