% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-seu_obj.R
\name{.generate_original_seu_metadata}
\alias{.generate_original_seu_metadata}
\title{generate original Seurat Object metadata}
\usage{
.generate_original_seu_metadata(raw_count_matrix, tissue_position_matrix)
}
\arguments{
\item{raw_count_matrix}{the count  matrix of raw dataset}

\item{tissue_position_matrix}{the matrix of tissue position}
}
\description{
generate original Seurat Object metadata
}
