% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-seu_obj.R
\name{.generate_seu_metadata_with_cluster_symbol}
\alias{.generate_seu_metadata_with_cluster_symbol}
\title{generate Seurat Object metadata with cluster symbol}
\usage{
.generate_seu_metadata_with_cluster_symbol(
  original_seu_metadata,
  in_tissue_barcode,
  raw_count_matrix
)
}
\arguments{
\item{original_seu_metadata}{original Seurat Object metadata}

\item{in_tissue_barcode}{the barcodes in tissue}

\item{raw_count_matrix}{the matrix of raw count dataset}
}
\description{
generate Seurat Object metadata with cluster symbol
}
