% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-seu_obj.R
\name{.generate_seu_metadata_with_hematoma_center_and_edge_symbol}
\alias{.generate_seu_metadata_with_hematoma_center_and_edge_symbol}
\title{add hematoma center and edge symbols to Seurat Object metadata}
\usage{
.generate_seu_metadata_with_hematoma_center_and_edge_symbol(hematoma)
}
\arguments{
\item{hematoma}{the Hematoma class}
}
\description{
add hematoma center and edge symbols to Seurat Object metadata
}
