% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-spatial_pattern.R
\name{.genereate_chunk_barcodes}
\alias{.genereate_chunk_barcodes}
\title{generate chunk barcodes}
\usage{
.genereate_chunk_barcodes(barcodes, chunk_bins)
}
\arguments{
\item{barcodes}{the barcode for chunk}

\item{chunk_bins}{the bins of chunk set}
}
\description{
generate chunk barcodes
}
