% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-spatial_pattern.R
\name{.genetate_distance_matrix}
\alias{.genetate_distance_matrix}
\title{generate the distance matrix of barcodes}
\usage{
.genetate_distance_matrix(
  seu_meta,
  spatial_chunks,
  kernel_function = "Gaussian"
)
}
\arguments{
\item{seu_meta}{the Seurat Object metadata}

\item{spatial_chunks}{the chunks for calculation}

\item{kernel_function}{choose the kernel function}
}
\description{
generate the distance matrix of barcodes
}
