% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-general.R
\name{.get_another_dataset_for_plotting}
\alias{.get_another_dataset_for_plotting}
\title{get another dataset for plotting}
\usage{
.get_another_dataset_for_plotting(ich_mouse)
}
\arguments{
\item{ich_mouse}{the class of ich_mouse}
}
\description{
get another dataset for plotting
}
