% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-reference_dataset.R
\name{.get_de_genes_from_log2FC_datatable}
\alias{.get_de_genes_from_log2FC_datatable}
\title{get differential exoression genes from log2FC data.table}
\usage{
.get_de_genes_from_log2FC_datatable(log2FC_dt)
}
\arguments{
\item{log2FC_dt}{the log2FC data.table}
}
\description{
get differential exoression genes from log2FC data.table
}
