% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-seu_obj.R
\name{.get_seu_obj_metadata}
\alias{.get_seu_obj_metadata}
\title{get the Seurat Object metadata}
\usage{
.get_seu_obj_metadata(seu_obj)
}
\arguments{
\item{seu_obj}{A Seurat Object}
}
\description{
get the Seurat Object metadata
}
