% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-reference_dataset.R
\name{.housekeeping_genes}
\alias{.housekeeping_genes}
\title{find housekeeping gene}
\usage{
.housekeeping_genes(reference_dataset)
}
\arguments{
\item{reference_dataset}{the reference dataset}
}
\description{
find housekeeping gene
}
