% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-identify_gene_distribution_pattern.R
\name{.identify_gene_distribution_pattern}
\alias{.identify_gene_distribution_pattern}
\title{identify gene distribution pattern}
\usage{
.identify_gene_distribution_pattern(
  threshold = 2,
  gene_set = character(),
  diff_expr_symbol = "edge-normal",
  ...
)

.identify_gene_distribution_pattern(
  threshold = 2,
  gene_set = character(),
  diff_expr_symbol = "edge-normal",
  ...
)
}
\arguments{
\item{threshold}{the avg_log2FC threshold}

\item{gene_set}{the gene set}

\item{ich_mouse}{the class of ICH_Mouse}
}
\description{
identify gene distribution pattern

identify gene distribution difference
}
