% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-reference_dataset.R
\name{.integrate_immune_mouse_RNA_seq_dataset}
\alias{.integrate_immune_mouse_RNA_seq_dataset}
\title{integrate immune mouse RNA seq dataset based on cell type}
\usage{
.integrate_immune_mouse_RNA_seq_dataset()
}
\description{
integrate immune mouse RNA seq dataset based on cell type
}
