% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-enrichment_analysis.R
\name{.integrate_iteration_cluster_results}
\alias{.integrate_iteration_cluster_results}
\title{integrate iteration cluster results}
\usage{
.integrate_iteration_cluster_results(GO_results)
}
\arguments{
\item{GO_results}{the results of GO enrichment}
}
\description{
integrate iteration cluster results
}
