% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-reference_dataset.R
\name{.integrate_mouse_RNA_seq_dataset}
\alias{.integrate_mouse_RNA_seq_dataset}
\title{integrate mouse RNA seq dataset based on cell type}
\usage{
.integrate_mouse_RNA_seq_dataset()
}
\description{
integrate mouse RNA seq dataset based on cell type
}
