% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-general.R
\name{.load_raw_count_matrix}
\alias{.load_raw_count_matrix}
\title{load raw count matrix}
\usage{
.load_raw_count_matrix(raw_count_matrix_address)
}
\arguments{
\item{raw_count_matrix_address}{the file address of raw count matrix}
}
\description{
load raw count matrix
}
