% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-general.R
\name{.load_tissue_position_matrix}
\alias{.load_tissue_position_matrix}
\title{load tissue position matrix}
\usage{
.load_tissue_position_matrix(tissue_position_address)
}
\arguments{
\item{tissue_position_address}{the address of tissue position matrix}
}
\description{
load tissue position matrix
}
