% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-graph.R
\name{.plotting_cnetplot}
\alias{.plotting_cnetplot}
\title{plotting cnetplot}
\usage{
.plotting_cnetplot(GO_results)
}
\arguments{
\item{GO_results}{the results of GO enrichment}
}
\description{
plotting cnetplot
}
