% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-general.R
\name{.plotting_GMM_barchart}
\alias{.plotting_GMM_barchart}
\title{plotting GMM barchart}
\usage{
.plotting_GMM_barchart(ich_mouse)
}
\description{
plotting GMM barchart
}
\section{Slots}{

\describe{
\item{\code{ich_mouse}}{the class of ICH_Mouse}
}}

