% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-graph.R
\name{.plotting_GO_term_cluster_heatmap}
\alias{.plotting_GO_term_cluster_heatmap}
\title{plotting GO term cluster heatmap}
\usage{
.plotting_GO_term_cluster_heatmap(ich_mouse, diff_expr_symbol = "edge-normal")
}
\arguments{
\item{ich_mouse}{the class of ICH_Mouse}
}
\description{
plotting GO term cluster heatmap
}
