% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-graph.R
\name{.plotting_GO_term_similarity_heatmap}
\alias{.plotting_GO_term_similarity_heatmap}
\title{plotting GO term similarity heatmap}
\usage{
.plotting_GO_term_similarity_heatmap(ich_mouse)
}
\arguments{
\item{ich_mouse}{the ICH_Mouse class}
}
\description{
plotting GO term similarity heatmap
}
