% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-graph_spatial_image.R
\name{.plotting_spatial_image}
\alias{.plotting_spatial_image}
\title{plotting spatial image}
\usage{
.plotting_spatial_image(ich_mouse)
}
\arguments{
\item{ich_mouse}{the class of ICH_Mouse}
}
\description{
plotting spatial image
}
