% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-general.R
\name{.plotting_umap_plot}
\alias{.plotting_umap_plot}
\title{plotting umap plots}
\usage{
.plotting_umap_plot(ich_mouse, diff_expr_genes_symbol = "edge-normal")
}
\arguments{
\item{ich_mouse}{the class of ICH_Mouse}
}
\description{
plotting umap plots
}
