% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-general.R
\name{.plotting_violin_plot}
\alias{.plotting_violin_plot}
\title{plotting violin plot}
\usage{
.plotting_violin_plot(ich_mouse)
}
\arguments{
\item{ich_mouse}{the class of ICH_Mouse}
}
\description{
plotting violin plot
}
