% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-general.R
\name{.plotting_volcano_plot}
\alias{.plotting_volcano_plot}
\title{plotting volcano plot}
\usage{
.plotting_volcano_plot(ich_mouse)
}
\arguments{
\item{ich_mouse}{the ICH_Mouse class}
}
\description{
plotting volcano plot
}
