% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-general.R
\name{.save_barcodes}
\alias{.save_barcodes}
\title{save barcodes}
\usage{
.save_barcodes(tissue_position_matrix, in_tissue)
}
\arguments{
\item{tissue_position_matrix}{the matrix of tissue position}

\item{in_tissue}{the location of barcodes value: TRUE or FALSE}
}
\description{
save barcodes
}
