% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-general.R
\name{.save_spatial_image}
\alias{.save_spatial_image}
\title{save spatial image}
\usage{
.save_spatial_image(saving_path, spatial_image, width, height)
}
\arguments{
\item{saving_path}{the path for saving}

\item{spatial_image}{the class of Spatial Image}
}
\description{
save spatial image
}
