% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-general.R
\name{.save_spatial_image_with_single_gene}
\alias{.save_spatial_image_with_single_gene}
\title{save spatial image with single gene}
\usage{
.save_spatial_image_with_single_gene(saving_path, spatial_image)
}
\arguments{
\item{saving_path}{the path for saving}

\item{spatial_image}{the class of Spatial Image}
}
\description{
save spatial image with single gene
}
