% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-enrichment_analysis.R
\name{.split_GO_result_genes}
\alias{.split_GO_result_genes}
\title{split GO result genes}
\usage{
.split_GO_result_genes(GO_result)
}
\arguments{
\item{GO_result}{the GO enrichment result}
}
\description{
split GO result genes
}
