% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-reference_dataset.R
\name{.the_log2FC_one_type_to_one_type}
\alias{.the_log2FC_one_type_to_one_type}
\title{calculate the log2FC one type to one type}
\usage{
.the_log2FC_one_type_to_one_type(reference_dataset)
}
\arguments{
\item{reference_dataset}{the reference dataset}
}
\description{
calculate the log2FC one type to one type
}
