% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-reference_dataset.R
\name{.transfer_ref_dt_to_mat}
\alias{.transfer_ref_dt_to_mat}
\title{transfer reference dataset to matrix}
\usage{
.transfer_ref_dt_to_mat(ref_dt)
}
\arguments{
\item{ref_dt}{the reference dataset in the form of data.table}
}
\description{
transfer reference dataset to matrix
}
