% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-reference_dataset.R
\name{.variable_genes}
\alias{.variable_genes}
\title{find variable gene}
\usage{
.variable_genes(reference_dataset)
}
\arguments{
\item{reference_dataset}{the reference dataset}
}
\description{
find variable gene
}
