% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-Enrichment_Set.R
\docType{class}
\name{Enrichment_Set-class}
\alias{Enrichment_Set-class}
\title{a class for storing enrichment result}
\description{
a class for storing enrichment result
}
\section{Slots}{

\describe{
\item{\code{GO_set}}{the GO term set for analysis}

\item{\code{GO_enrich}}{the result of GO enrich}

\item{\code{KEGG_enrich}}{the result of KEGG enrich}

\item{\code{gene_information}}{the gene information}
}}

