% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-general.R
\name{export_data.table_as_excel}
\alias{export_data.table_as_excel}
\title{export data.table as excel}
\usage{
export_data.table_as_excel(data.table_obj, saving_path, file_name)
}
\arguments{
\item{data.table_obj}{a data.table object or list}

\item{saving_path}{the saving path}

\item{file_name}{the workbook name}
}
\description{
export data.table as excel
}
