% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-Save_Load_ICHMouse.R
\name{export_plotting_dataset,Graph_Database,character-method}
\alias{export_plotting_dataset,Graph_Database,character-method}
\title{export graph database}
\usage{
\S4method{export_plotting_dataset}{Graph_Database,character}(plotting_dataset, saving_path)
}
\arguments{
\item{plotting_dataset}{the Graph_Database class}

\item{saving_path}{the path for saving Spatial_Image}
}
\description{
export graph database
}
