% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-Save_Load_ICHMouse.R
\name{export_plotting_dataset,Spatial_Image,character-method}
\alias{export_plotting_dataset,Spatial_Image,character-method}
\title{export spatial image dataset}
\usage{
\S4method{export_plotting_dataset}{Spatial_Image,character}(plotting_dataset, saving_path)
}
\arguments{
\item{plotting_dataset}{the Spatial_Image class}

\item{saving_path}{the path for saving Spatial_Image}
}
\description{
export spatial image dataset
}
