% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-ICH_Mouse.R
\name{find_differential_expression_genes,ICH_Mouse-method}
\alias{find_differential_expression_genes,ICH_Mouse-method}
\title{find differential expression genes}
\usage{
\S4method{find_differential_expression_genes}{ICH_Mouse}(
  ich_mouse,
  cluster_symbol = c(3, 1),
  gene_set_name = "edge-normal"
)
}
\arguments{
\item{ich_mouse}{the ICH_Mouse class}

\item{cluster_symbol}{the cluster to contrast}

\item{gene_set_name}{the gene set to find expression genes}
}
\description{
find differential expression genes
}
