% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-Gene_Spatial_Pattern.R
\docType{class}
\name{Gene_Spatial_Pattern-class}
\alias{Gene_Spatial_Pattern-class}
\title{the Gene_Spatial_Pattern class}
\description{
the Gene_Spatial_Pattern class
}
\section{Slots}{

\describe{
\item{\code{pattern_name}}{the spatial pattern name}

\item{\code{gene_list}}{the gene list for classification}

\item{\code{spatial_chunk}}{the class of Spatial_Chunk}

\item{\code{characteristic_value_matrix}}{the characteristic value matrix}
}}

