% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-Graph_Database.R
\docType{class}
\name{Graph_Database-class}
\alias{Graph_Database-class}
\title{the class of Graph_Database}
\description{
the class of Graph_Database
}
\section{Slots}{

\describe{
\item{\code{graph_database}}{the graph database}
}}

