% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-Hematoma.R
\docType{class}
\name{Hematoma-class}
\alias{Hematoma-class}
\title{Hematoma: A class for identify the hematoma from spatial trascriptome dataset}
\description{
Hematoma: A class for identify the hematoma from spatial trascriptome dataset
}
\section{Slots}{

\describe{
\item{\code{analysis_symbol}}{a symbol of sample}

\item{\code{file_address}}{file address of dataset used in analysis}

\item{\code{tissue_position_matrix}}{save tissue position matrix of dataset}

\item{\code{color_set}}{a set of code for color}

\item{\code{raw_count_matrix}}{the raw count matrix from dataset}

\item{\code{original_seu_metadata}}{the original Seurat Object metadata}

\item{\code{seu_metadata_with_cluster_symbol}}{Seurat Object metadata with cluster symbol}

\item{\code{spatial_image}}{the spatial image with cluster}

\item{\code{identification_symbols}}{the symbols identified by user as the the symbol of hematoma}

\item{\code{giotto_instruction}}{the instruction of Giotto Object}

\item{\code{filtered_genes}}{the genes that been filtered}

\item{\code{filtered_barcodes}}{the filtered barcodes}
}}

