% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-Hematoma.R
\name{identify_hematoma,Hematoma-method}
\alias{identify_hematoma,Hematoma-method}
\title{identify the exact hematoma location}
\usage{
\S4method{identify_hematoma}{Hematoma}(hematoma)
}
\description{
@param hematoma the class of HematomaICHMousewch
}
