% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-ICH_Mouse.R
\name{identify_hematoma,Hematoma,numeric-method}
\alias{identify_hematoma,Hematoma,numeric-method}
\title{identify the exact hematoma location}
\usage{
\S4method{identify_hematoma}{Hematoma,numeric}(hematoma, hematoma_symbols)
}
\description{
@param hematoma the class of HematomaICHMousewch
@param hematoma_symbols the symbols identified by user as the the symbol of hematoma
}
