% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-Hematoma.R
\name{identify_hematoma}
\alias{identify_hematoma}
\title{the generic function of identify_hematoma}
\usage{
identify_hematoma(hematoma)
}
\arguments{
\item{hematoma}{the class of HematomaICHMousewch}
}
\description{
the generic function of identify_hematoma
}
