% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-Hematoma.R
\name{identify_hematoma_center_and_edge,Hematoma-method}
\alias{identify_hematoma_center_and_edge,Hematoma-method}
\title{identify the exact hematoma center and edge}
\usage{
\S4method{identify_hematoma_center_and_edge}{Hematoma}(hematoma)
}
\description{
@param hematoma the class of HematomaICHMousewch
}
