% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-Hematoma.R
\name{identify_hematoma_center_and_edge,Hematoma,numeric-method}
\alias{identify_hematoma_center_and_edge,Hematoma,numeric-method}
\title{identify the exact hematoma center and edge}
\usage{
\S4method{identify_hematoma_center_and_edge}{Hematoma,numeric}(hematoma, center_symbols)
}
\description{
@param hematoma the class of HematomaICHMousewch
@param center_symbols the symbols identified by user as the the symbol of hematoma center
}
