% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-Hematoma.R
\name{identify_hematoma_center_and_edge}
\alias{identify_hematoma_center_and_edge}
\title{identify hematoma center and edge}
\usage{
identify_hematoma_center_and_edge(hematoma)
}
\arguments{
\item{hematoma}{the class of HematomaICHMousewch}
}
\description{
identify hematoma center and edge
}
