% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-Gene_Spatial_Pattern.R
\name{initialize,Gene_Spatial_Pattern-method}
\alias{initialize,Gene_Spatial_Pattern-method}
\title{initialize class of Gene_Spatial_Pattern}
\usage{
\S4method{initialize}{Gene_Spatial_Pattern}(.Object, ich_mouse, gene_ls, pattern_na, bin_scale = 5000)
}
\arguments{
\item{ich_mouse}{the class of ich_mouse}

\item{gene_ls}{the genes for classfication}

\item{pattern_na}{the spatial pattern name}

\item{bin_scale}{the scale of the bin}
}
\description{
initialize class of Gene_Spatial_Pattern
}
