% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-Plotting_Dataset.R
\name{initialize,Plotting_Dataset-method}
\alias{initialize,Plotting_Dataset-method}
\title{initialize Plotting_Dataset}
\usage{
\S4method{initialize}{Plotting_Dataset}(.Object, ich_mouse, initialization)
}
\arguments{
\item{ich_mouse}{the class of ICH_Mouse}
}
\description{
initialize Plotting_Dataset
}
