% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-Spatial_Chunk.R
\name{initialize,Spatial_Chunk-method}
\alias{initialize,Spatial_Chunk-method}
\title{initialize the class of Spatial_Chunk}
\usage{
\S4method{initialize}{Spatial_Chunk}(.Object, barcodes, chunk_set_name, bin_scale = 5000)
}
\arguments{
\item{barcodes}{the barcodes for chunking}

\item{chunk_set_name}{the name of the chunk set}

\item{bin_scale}{the scale of the chunk}
}
\description{
initialize the class of Spatial_Chunk
}
