% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-Spatial_Image.R
\name{initialize,Spatial_Image-method}
\alias{initialize,Spatial_Image-method}
\title{Initialize the class of Spatial_Image}
\usage{
\S4method{initialize}{Spatial_Image}(.Object, image_set_name, spatial_image_ls, initialization)
}
\arguments{
\item{image_set_name}{the name of image set}

\item{spatial_image_ls}{a list contain a set of spatial image}
}
\description{
Initialize the class of Spatial_Image
}
