% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-Save_Load_ICHMouse.R
\name{load_graph,ANY-method}
\alias{load_graph,ANY-method}
\title{load to Graph_Database}
\usage{
\S4method{load_graph}{ANY}(graph_name, data_symbol = "ich_mouse", loading_path = NULL)
}
\description{
load to Graph_Database
}
