% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-Hematoma.R
\name{load_Hematoma,character-method}
\alias{load_Hematoma,character-method}
\title{load Hematoma from local}
\usage{
\S4method{load_Hematoma}{character}(loading_path)
}
\arguments{
\item{loading_path}{the path saving Hematoma}
}
\description{
load Hematoma from local
}
