% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-Hematoma.R
\name{load_Hematoma}
\alias{load_Hematoma}
\title{load Hematoma from local}
\usage{
load_Hematoma(loading_path)
}
\arguments{
\item{loading_path}{the path saving Hematoma}
}
\description{
load Hematoma from local
}
