% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-ICH_Mouse.R
\name{load_ICH_Mouse,character-method}
\alias{load_ICH_Mouse,character-method}
\title{load ICH_Mouse from local}
\usage{
\S4method{load_ICH_Mouse}{character}(loading_path)
}
\arguments{
\item{loading_path}{the path saving ICH_Mouse}
}
\description{
load ICH_Mouse from local
}
