% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-Save_Load_ICHMouse.R
\name{Load_ICHMouse,ANY-method}
\alias{Load_ICHMouse,ANY-method}
\title{load Hematoma}
\usage{
\S4method{Load_ICHMouse}{ANY}(class_symbol, data_symbol = character(), loading_path = NULL)

\S4method{Load_ICHMouse}{ANY}(class_symbol, data_symbol = character(), loading_path = NULL)
}
\arguments{
\item{class_symbol}{the class type}

\item{data_symbol}{the data symbol}

\item{loading_path}{the path loading data}
}
\description{
load Hematoma

load ICH_Mouse
}
