% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-Plotting_Dataset.R
\name{load_Plotting_Dataset,character-method}
\alias{load_Plotting_Dataset,character-method}
\title{load Plotting_Dataset from local}
\usage{
\S4method{load_Plotting_Dataset}{character}(loading_path)
}
\arguments{
\item{loading_path}{the path saving Plotting_Dataset}
}
\description{
load Plotting_Dataset from local
}
