% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-Plotting_Dataset.R
\name{load_Plotting_Dataset}
\alias{load_Plotting_Dataset}
\title{load Plotting_Dataset from local}
\usage{
load_Plotting_Dataset(loading_path)
}
\arguments{
\item{loading_path}{the path saving Plotting_Dataset}
}
\description{
load Plotting_Dataset from local
}
