% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-Plotting_Dataset.R
\docType{class}
\name{Plotting_Dataset-class}
\alias{Plotting_Dataset-class}
\title{the class for Plotting}
\description{
the class for Plotting
}
\section{Slots}{

\describe{
\item{\code{spatial_image}}{the spatial image}

\item{\code{spatial_single_gene_image}}{the spatial single gene image}

\item{\code{violin_plot}}{the violin plot}

\item{\code{volcano_plot}}{the volcano plot}

\item{\code{heatmap}}{the heatmap}

\item{\code{bubble_chart}}{the bubble chart}
}}

