% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-graph.R
\name{save_gtable_plot}
\alias{save_gtable_plot}
\title{save gtable plot}
\usage{
save_gtable_plot(gtable_plot, saving_path, file_name)
}
\arguments{
\item{gtable_plot}{a plot in the form of gtable}

\item{saving_path}{the path for saving plot}

\item{file_name}{the file name}
}
\description{
save gtable plot
}
