% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-ICH_Mouse.R
\name{save_Hematoma,Hematoma,character-method}
\alias{save_Hematoma,Hematoma,character-method}
\title{save Hematoma class}
\usage{
\S4method{save_Hematoma}{Hematoma,character}(hematoma, saving_path)
}
\arguments{
\item{hematoma}{the Hematoma class}

\item{saving_path}{the path to save}
}
\description{
save Hematoma class
}
