% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-Hematoma.R
\name{save_Hematoma}
\alias{save_Hematoma}
\title{save Hematoma class}
\usage{
save_Hematoma(hematoma, saving_path)
}
\arguments{
\item{hematoma}{the Hematoma class}

\item{saving_path}{the path to save}
}
\description{
save Hematoma class
}
