% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-ICH_Mouse.R
\name{save_ICH_Mouse,ICH_Mouse,character-method}
\alias{save_ICH_Mouse,ICH_Mouse,character-method}
\title{save ICH_Mouse class}
\usage{
\S4method{save_ICH_Mouse}{ICH_Mouse,character}(ich_mouse, saving_path)
}
\arguments{
\item{ich_mouse}{the ICH_Mouse class}

\item{saving_path}{the path to save}
}
\description{
save ICH_Mouse class
}
