% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-ICH_Mouse.R
\name{save_ICH_Mouse}
\alias{save_ICH_Mouse}
\title{save ICH_Mouse class}
\usage{
save_ICH_Mouse(ich_mouse, saving_path)
}
\arguments{
\item{ich_mouse}{the ICH_Mouse class}

\item{saving_path}{the path to save}
}
\description{
save ICH_Mouse class
}
