% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-Save_ICHMouse.R
\name{Save_ICHMouse,Hematoma,character-method}
\alias{Save_ICHMouse,Hematoma,character-method}
\title{save Hematoma}
\usage{
\S4method{Save_ICHMouse}{Hematoma,character}(Class_ICHMouse, saving_path = getwd())
}
\arguments{
\item{Class_ICHMouse}{the class of ICHMouse}

\item{saving_path}{the path for saving}
}
\description{
save Hematoma
}
