% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-Save_Load_ICHMouse.R
\name{Save_ICHMouse,Hematoma-method}
\alias{Save_ICHMouse,Hematoma-method}
\title{save Hematoma}
\usage{
\S4method{Save_ICHMouse}{Hematoma}(Class_ICHMouse, data_symbol = "ich_mouse", saving_path = NULL)
}
\arguments{
\item{Class_ICHMouse}{the class of ICHMouse}

\item{data_symbol}{the saving symbol}

\item{saving_path}{the path for saving}
}
\description{
save Hematoma
}
