% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-Save_Load_ICHMouse.R
\name{Save_ICHMouse,Spatial_Image-method}
\alias{Save_ICHMouse,Spatial_Image-method}
\title{save Spatial_Image}
\usage{
\S4method{Save_ICHMouse}{Spatial_Image}(Class_ICHMouse, data_symbol = "ich_mouse", saving_path = NULL)
}
\arguments{
\item{Class_ICHMouse}{the class of ICHMouse}

\item{data_symbol}{the saving symbol}

\item{saving_path}{the path for saving}
}
\description{
save Spatial_Image
}
