% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-Plotting_Dataset.R
\name{save_Plotting_Dataset,Plotting_Dataset,character-method}
\alias{save_Plotting_Dataset,Plotting_Dataset,character-method}
\title{save Plotting_Dataset class}
\usage{
\S4method{save_Plotting_Dataset}{Plotting_Dataset,character}(plotting_dataset, saving_path)
}
\arguments{
\item{plotting_dataset}{the Plotting_Dataset class}

\item{saving_path}{the path to save}
}
\description{
save Plotting_Dataset class
}
