% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-Plotting_Dataset.R
\name{save_Plotting_Dataset}
\alias{save_Plotting_Dataset}
\title{save Plotting_Dataset class}
\usage{
save_Plotting_Dataset(plotting_dataset, saving_path)
}
\arguments{
\item{plotting_dataset}{the Plotting_Dataset class}

\item{saving_path}{the path to save}
}
\description{
save Plotting_Dataset class
}
