% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-Spatial_Image.R
\name{save_spatial_image}
\alias{save_spatial_image}
\title{save spatial image}
\usage{
save_spatial_image(spatial_image, saving_path, height, width)
}
\arguments{
\item{spatial_image}{the class of Spatial Image}

\item{saving_path}{the path for saving}
}
\description{
save spatial image
}
