% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-ICH_Mouse.R
\name{save_spatial_image_with_single_gene,ICH_Mouse,character,character-method}
\alias{save_spatial_image_with_single_gene,ICH_Mouse,character,character-method}
\title{save spatial image with single gene}
\usage{
\S4method{save_spatial_image_with_single_gene}{ICH_Mouse,character,character}(ich_mouse, saving_path, image_set)
}
\arguments{
\item{ich_mouse}{the class of ICH_Mouse}

\item{saving_path}{the path for saving}

\item{image_set}{the image set to save}
}
\description{
save spatial image with single gene
}
