% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-Spatial_Image.R
\name{save_spatial_image_with_single_gene,Spatial_Image,character-method}
\alias{save_spatial_image_with_single_gene,Spatial_Image,character-method}
\title{save spatial image with single gene}
\usage{
\S4method{save_spatial_image_with_single_gene}{Spatial_Image,character}(spatial_image, saving_path)
}
\arguments{
\item{spatial_image}{the class of Spatial Image}

\item{saving_path}{the path for saving}
}
\description{
save spatial image with single gene
}
