% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-Hematoma.R
\name{show_image,list-method}
\alias{show_image,list-method}
\title{show ggplotGrob image}
\usage{
\S4method{show_image}{list}(ggplot_image)
}
\arguments{
\item{ggplot_image}{a ggplot image in the form of gtable}
}
\description{
show ggplotGrob image
}
