% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-Hematoma.R
\name{show_image}
\alias{show_image}
\title{show ggplotGrob image}
\usage{
show_image(ggplot_image)
}
\arguments{
\item{ggplot_image}{a ggplot image in the form of gtable}
}
\description{
show ggplotGrob image
}
