% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-Spatial_Chunk.R
\docType{class}
\name{Spatial_Chunk-class}
\alias{Spatial_Chunk-class}
\title{the class Spatial_Chunk}
\description{
the class Spatial_Chunk
}
\section{Slots}{

\describe{
\item{\code{chunk_bin}}{the bins of the chunk}

\item{\code{chunk_barcode}}{the barcode of the chunk}

\item{\code{chunk_parameter}}{the parameter of the chunk}
}}

