% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-Spatial_Image.R
\docType{class}
\name{Spatial_Image-class}
\alias{Spatial_Image-class}
\title{the class fot save Giottto Spatial Image}
\description{
the class fot save Giottto Spatial Image
}
\section{Slots}{

\describe{
\item{\code{image_set_name}}{the name of image set}

\item{\code{gene_name}}{a gene list for plotting the spatial image}

\item{\code{spatial_image}}{a list of spatial image}
}}

