% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-Enrichment_Set.R
\name{add_GO_term_set,Enrichment_Set,character,character-method}
\alias{add_GO_term_set,Enrichment_Set,character,character-method}
\title{add GO term set}
\usage{
\S4method{add_GO_term_set}{Enrichment_Set,character,character}(enrichment_set, GO_term_set, GO_set_name, ich_mouse)
}
\arguments{
\item{enrichment_set}{the class of Enrichment_Set}

\item{GO_term_set}{the set of GO ID}

\item{GO_set_name}{the name of the GO set}

\item{ich_mouse}{the class of ICH_Mouse}
}
\description{
add GO term set
}
