% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-Enrichment_Set.R
\name{add_GO_term_set}
\alias{add_GO_term_set}
\title{add GO term set}
\usage{
add_GO_term_set(
  enrichment_set,
  GO_term_set_ls,
  ich_mouse,
  diff_symbol = "edge-normal"
)
}
\arguments{
\item{enrichment_set}{the class of Enrichment_Set}

\item{GO_term_set_ls}{the set of GO ID}

\item{ich_mouse}{the class of ICH_Mouse}
}
\description{
add GO term set
}
