% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-Hematoma.R
\name{Create_Hematoma}
\alias{Create_Hematoma}
\title{Hematoma constructor}
\usage{
Create_Hematoma(
  analysis_symbol,
  raw_count_matrix_address,
  filtered_count_matrix_address,
  tissue_position_address,
  background_image_address,
  giotto_python_path,
  giotto_results_folder = "giotto_results",
  initialization = TRUE
)
}
\arguments{
\item{analysis_symbol}{a symbol of sample}

\item{raw_count_matrix_address}{address for raw count matrix}

\item{tissue_position_address}{address for tissue position matrix}

\item{background_image_address}{address for background address}

\item{initialization}{whether start at beginning}

\item{filtered_count_matrix}{address for filtered count matrix}
}
\description{
Hematoma constructor
}
