% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-ICH_Mouse.R
\name{Create_ICH_Mouse}
\alias{Create_ICH_Mouse}
\title{ICH_Mouse constructor}
\usage{
Create_ICH_Mouse(
  analysis_symbol,
  raw_count_matrix_address,
  filtered_count_matrix_address,
  tissue_position_address,
  background_image_address,
  giotto_python_path,
  giotto_results_folder,
  hematoma_symbols,
  center_symbols,
  initialization = TRUE,
  spaceranger_umap_address = character(),
  spaceranger_cluster_address = character()
)
}
\arguments{
\item{analysis_symbol}{a symbol of sample}

\item{raw_count_matrix_address}{address for raw count matrix}

\item{tissue_position_address}{address for tissue position matrix}

\item{background_image_address}{address for background address}

\item{hematoma_symbols}{the symbols identified by user as the the symbol of hematoma}

\item{center_symbols}{the symbols identified by user as the the symbol of hematoma center}

\item{initialization}{whether start at beginning}

\item{filtered_count_matrix}{address for filtered count matrix}
}
\description{
ICH_Mouse constructor
}
