% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-Plotting_Dataset.R
\name{Create_Plotting_Dataset}
\alias{Create_Plotting_Dataset}
\title{constructor of Plotting_Dataset}
\usage{
Create_Plotting_Dataset(ich_mouse, initialization = TRUE)
}
\arguments{
\item{ich_mouse}{the class of ICH_Mouse}
}
\description{
constructor of Plotting_Dataset
}
