% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-ICH_Mouse.R
\name{create_single_gene_spatial_image}
\alias{create_single_gene_spatial_image}
\title{create single gene spatial image}
\usage{
create_single_gene_spatial_image(ich_mouse, gene_ls)
}
\arguments{
\item{ich_mouse}{the class of ICH_Mouse}

\item{gene_ls}{the gene list to create spatial image}
}
\description{
create single gene spatial image
}
