% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-graph.R
\name{.create_count_distribution_map}
\alias{.create_count_distribution_map}
\title{create count distribution map}
\usage{
.create_count_distribution_map(seu_meta)
}
\arguments{
\item{seu_meta}{the Seurat metadata}
}
\description{
create count distribution map
}
