% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-graph.R
\name{.create_gene_distribution_map}
\alias{.create_gene_distribution_map}
\title{create gene distribution map}
\usage{
.create_gene_distribution_map(
  ich_mouse,
  diff_expr_gene_symbol = "edge-normal",
  aim_gene = character(),
  background_genes = c("Hbb-bt", "Hbb-bs", "Hba-a2"),
  point_size = 1e-04,
  normalized_data = FALSE,
  scaled_data = TRUE
)
}
\arguments{
\item{ich_mouse}{the class of ICH_Mouse}

\item{diff_expr_gene_symbol}{the differential expression genes symbol}

\item{aim_gene}{the aim gene for plotting}

\item{background_genes}{the background genes for graph}
}
\description{
create gene distribution map
}
