% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-graph_spatial_image.R
\name{.create_gene_distribution_map}
\alias{.create_gene_distribution_map}
\title{create gene distribution map}
\usage{
.create_gene_distribution_map(
  ich_mouse,
  diff_expr_gene_symbol = "edge-normal",
  aim_gene = character(),
  background_genes = character(),
  point_size = 1e-04,
  scaled_data = TRUE,
  legend_position = "right",
  output_dataset = FALSE
)
}
\arguments{
\item{ich_mouse}{the class of ICH_Mouse}

\item{diff_expr_gene_symbol}{the differential expression genes symbol}

\item{aim_gene}{the aim gene for plotting}

\item{background_genes}{the background genes for graph}

\item{output_dataset}{whether to output plotting dataset}
}
\description{
create gene distribution map
}
