% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-graph.R
\name{.create_gene_distribution_map}
\alias{.create_gene_distribution_map}
\title{create gene distribution map}
\usage{
.create_gene_distribution_map(
  seu_meta,
  raw_count_matrix,
  filtered_genes,
  filtered_barcodes,
  diff_expr_gene,
  aim_gene = character(),
  background_genes = c("Hbb-bt", "Hbb-bs", "Hba-a2"),
  point_size = 1e-04
)
}
\arguments{
\item{seu_meta}{the Seurat metadata}

\item{raw_count_matrix}{the raw count matrix}

\item{filtered_genes}{the filtered genes}

\item{filtered_barcodes}{the filtered barcodes}

\item{diff_expr_gene}{the differential expression genes}

\item{aim_gene}{the aim gene for plotting}

\item{background_genes}{the background genes for graph}
}
\description{
create gene distribution map
}
