% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-graph_spatial_image.R
\name{.create_percent_distribution_map}
\alias{.create_percent_distribution_map}
\title{create percent distribution map}
\usage{
.create_percent_distribution_map(
  ich_mouse,
  aim_gene,
  diff_expr_gene_symbol = "edge-normal",
  distribution_region = "center_edge_symbol"
)
}
\arguments{
\item{ich_mouse}{the class of ICH_Mouse}

\item{aim_gene}{the aim gene list}

\item{distribution_region}{the region for distribution}
}
\description{
create percent distribution map
}
