% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-graph.R
\name{.create_spatial_image_with_cluster_symbol}
\alias{.create_spatial_image_with_cluster_symbol}
\title{create spatial image with cluster symbol}
\usage{
.create_spatial_image_with_cluster_symbol(
  ich_mouse,
  cluster_symbol,
  self_definition_color = character(),
  theme_param = list(),
  plot_title = character(),
  legend_lable = character(),
  show_image = TRUE,
  show_legend_label = FALSE,
  show_plot_title = FALSE
)
}
\arguments{
\item{ich_mouse}{the ICH_Mouse class}

\item{cluster_symbol}{the cluster for spatial image}

\item{self_definition_color}{a set of color defined by user}

\item{plot_title}{the title of spatial image}

\item{legend_lable}{the legend label}
}
\description{
create spatial image with cluster symbol
}
