% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-graph.R
\name{.create_spatial_image_with_cluster_symbol}
\alias{.create_spatial_image_with_cluster_symbol}
\title{create spatial image with cluster symbol}
\usage{
.create_spatial_image_with_cluster_symbol(
  in_tissue_metadata,
  cluster_symbol,
  raw_count_matrix,
  background_image_address,
  giotto_instruction,
  self_definition_color = character(),
  theme_param = list(),
  plot_title = character(),
  legend_lable = character(),
  show_image = TRUE,
  show_legend_label = FALSE,
  show_plot_title = FALSE
)
}
\arguments{
\item{in_tissue_metadata}{the Seurat Object metadata of in tissue barcodes}

\item{cluster_symbol}{the cluster for spatial image}

\item{raw_count_matrix}{the matrix of raw count dataset}

\item{background_image_address}{image for background}

\item{giotto_instruction}{the instruction of Giotto Object}

\item{self_definition_color}{a set of color defined by user}
}
\description{
create spatial image with cluster symbol
}
