% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-graph.R
\name{.create_spatial_image_with_single_gene}
\alias{.create_spatial_image_with_single_gene}
\title{create spatial image with single gene}
\usage{
.create_spatial_image_with_single_gene(
  seu_metadata_with_cluster_symbol,
  gene_ls,
  raw_count_matrix,
  background_image_address,
  giotto_instruction
)
}
\arguments{
\item{seu_metadata_with_cluster_symbol}{the Seurat Object metadata of in tissue barcodes}

\item{gene_ls}{the gene list for creating spatial image}

\item{raw_count_matrix}{the matrix of raw count dataset}

\item{background_image_address}{image for background}

\item{giotto_instruction}{the instruction of Giotto Object}
}
\description{
create spatial image with single gene
}
