% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-graph.R
\name{.create_umap_plot}
\alias{.create_umap_plot}
\title{create umap plot}
\usage{
.create_umap_plot(ich_mouse, spaceranger_result)
}
\arguments{
\item{ich_mouse}{the class of ICH_Mouse}

\item{spaceranger_result}{the file address of spaceranger_result}
}
\description{
create umap plot
}
