% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-graph.R
\name{.create_Venn_Diagram}
\alias{.create_Venn_Diagram}
\title{create Venn Diagram}
\usage{
.create_Venn_Diagram(data_list)
}
\arguments{
\item{data_list}{the data list for Venn Diagram}
}
\description{
create Venn Diagram
}
