% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-general.R
\name{.find_filtered_genes}
\alias{.find_filtered_genes}
\title{find filtered genes}
\usage{
.find_filtered_genes(raw_count_matrix, original_seu_metadata)
}
\arguments{
\item{raw_count_matrix}{the matrix of raw count dataset}

\item{original_seu_metadata}{the original Seurat Object metadata}
}
\description{
find filtered genes
}
