% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-graph.R
\name{.generate_bubble_chart_y_tick}
\alias{.generate_bubble_chart_y_tick}
\title{generate bubble chart y tick}
\usage{
.generate_bubble_chart_y_tick(plotting_dataset)
}
\arguments{
\item{plotting_dataset}{the plotting dataset}
}
\description{
generate bubble chart y tick
}
