% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-graph.R
\name{.generate_bubble_dataset}
\alias{.generate_bubble_dataset}
\title{generate bubble dataset}
\usage{
.generate_bubble_dataset(gene_info, GO_info, gene_num)
}
\arguments{
\item{gene_info}{the gene information}

\item{GO_info}{the GO term information}
}
\description{
generate bubble dataset
}
