% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-enrichment_analysis.R
\name{.generate_single_gene_GO_infomation}
\alias{.generate_single_gene_GO_infomation}
\title{generate single gene GO information}
\usage{
.generate_single_gene_GO_infomation(enrichment_set, ich_mouse, GO_term_set_ls)
}
\arguments{
\item{enrichment_set}{the class of Enrichment_Set}

\item{ich_mouse}{the class of ICH_Mouse}

\item{GO_term_set_ls}{the set of GO ID}
}
\description{
generate single gene GO information
}
