% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-identify_gene_distribution_pattern.R
\name{.identify_gene_distribution_pattern}
\alias{.identify_gene_distribution_pattern}
\title{identify gene distribution pattern}
\usage{
.identify_gene_distribution_pattern(ich_mouse, threshold)
}
\arguments{
\item{ich_mouse}{the class of ICH_Mouse}

\item{threshold}{the avg_log2FC threshold}
}
\description{
identify gene distribution pattern
}
