% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-general.R
\name{.integrate_file_address}
\alias{.integrate_file_address}
\title{integrate file address}
\usage{
.integrate_file_address(
  raw_count_matrix_address,
  filtered_count_matrix_address,
  tissue_position_address,
  background_image_address
)
}
\arguments{
\item{raw_count_matrix_address}{address for raw count matrix}

\item{tissue_position_address}{address for tissue position matrix}

\item{background_image_address}{address for background address}

\item{filtered_count_matrix}{address for filtered count matrix}
}
\description{
integrate file address
}
