% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-general.R
\name{.plotting_bubble_chart}
\alias{.plotting_bubble_chart}
\title{plotting bubble chart}
\usage{
.plotting_bubble_chart(ich_mouse, GO_ID_ls)
}
\arguments{
\item{ich_mouse}{the ICH_Mouse class}

\item{GO_ID_ls}{a list of GO ID for plotting}
}
\description{
plotting bubble chart
}
