% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-general.R
\name{.plotting_bubble_chart}
\alias{.plotting_bubble_chart}
\title{plotting bubble chart}
\usage{
.plotting_bubble_chart(ich_mouse, gene_num = 10, group_na = character())
}
\arguments{
\item{ich_mouse}{the ICH_Mouse class}

\item{group_na}{the group name}
}
\description{
plotting bubble chart
}
