% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-general.R
\name{.save_spatial_image_with_single_gene}
\alias{.save_spatial_image_with_single_gene}
\title{save spatial image with single gene}
\usage{
.save_spatial_image_with_single_gene(ich_mouse, saving_path, image_set)
}
\arguments{
\item{ich_mouse}{the class of ICH_Mouse}

\item{saving_path}{the path for saving}

\item{image_set}{the image set to save}
}
\description{
save spatial image with single gene
}
