% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-reference_dataset.R
\name{.the_log2FC_one_type_to_other_type}
\alias{.the_log2FC_one_type_to_other_type}
\title{calculate the log2FC one type to other type}
\usage{
.the_log2FC_one_type_to_other_type()
}
\description{
calculate the log2FC one type to other type
}
