% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-Plotting_Dataset.R
\name{export_plotting_dataset}
\alias{export_plotting_dataset}
\title{export plotting dataset}
\usage{
export_plotting_dataset(plotting_dataset, saving_path)
}
\arguments{
\item{plotting_dataset}{the Plotting_Dataset class}

\item{saving_path}{the path for saving plotting_dataset}
}
\description{
export plotting dataset
}
