% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-ICH_Mouse.R
\name{find_differential_expression_genes,ICH_Mouse,numeric,character-method}
\alias{find_differential_expression_genes,ICH_Mouse,numeric,character-method}
\title{find differential expression genes}
\usage{
\S4method{find_differential_expression_genes}{ICH_Mouse,numeric,character}(ich_mouse, cluster_symbol, gene_set_name)
}
\arguments{
\item{ich_mouse}{the ICH_Mouse class}

\item{cluster_symbol}{the cluster to contrast}
}
\description{
find differential expression genes
}
