% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-ICH_Mouse.R
\name{find_differential_expression_genes}
\alias{find_differential_expression_genes}
\title{find differential expression genes}
\usage{
find_differential_expression_genes(ich_mouse, cluster_symbol, gene_set_name)
}
\arguments{
\item{ich_mouse}{the ICH_Mouse class}

\item{cluster_symbol}{the cluster to contrast}
}
\description{
find differential expression genes
}
