% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-ICH_Mouse.R
\docType{class}
\name{ICH_Mouse-class}
\alias{ICH_Mouse-class}
\title{ICH_Mouse a class for analysis ICH_Mouse}
\description{
ICH_Mouse a class for analysis ICH_Mouse
}
\section{Slots}{

\describe{
\item{\code{diff_expr_genes}}{the differential expression genes}

\item{\code{symbol_genes}}{the symbol genes}

\item{\code{GO_enrichment}}{the result of GO_enrichment}

\item{\code{GO_cluster}}{the cluster of GO term}

\item{\code{GO_ID_group}}{the aimed analysis GO ID group}

\item{\code{gene_group}}{the gene group}
}}

