% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-ICH_Mouse.R
\name{identify_hematoma}
\alias{identify_hematoma}
\title{the generic function of identify_hematoma}
\usage{
identify_hematoma(hematoma, hematoma_symbols)
}
\arguments{
\item{hematoma}{the class of HematomaICHMousewch}

\item{hematoma_symbols}{the symbols identified by user as the the symbol of hematoma}
}
\description{
the generic function of identify_hematoma
}
