% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-ICH_Mouse.R
\name{identify_hematoma_center_and_edge}
\alias{identify_hematoma_center_and_edge}
\title{identify hematoma center and edge}
\usage{
identify_hematoma_center_and_edge(hematoma, center_symbols)
}
\arguments{
\item{hematoma}{the class of HematomaICHMousewch}

\item{center_symbols}{the symbols identified by user as the the symbol of hematoma center}
}
\description{
identify hematoma center and edge
}
