% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-Enrichment_Set.R
\name{initialize,Enrichment_Set-method}
\alias{initialize,Enrichment_Set-method}
\title{initialize Enrichment_Set}
\usage{
\S4method{initialize}{Enrichment_Set}(.Object, gene_list, ich_mouse)
}
\arguments{
\item{gene_list}{the gene list to conduct enrichment analysis}

\item{ich_mouse}{the ICH_Mouse class}
}
\description{
initialize Enrichment_Set
}
