% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-ICH_Mouse.R
\name{initialize,Hematoma-method}
\alias{initialize,Hematoma-method}
\title{Initialize class Hematoma}
\usage{
\S4method{initialize}{Hematoma}(
  .Object,
  analysis_symbol,
  raw_count_matrix_address,
  filtered_count_matrix_address,
  tissue_position_address,
  background_image_address,
  giotto_python_path,
  giotto_results_folder
)
}
\arguments{
\item{analysis_symbol}{a symbol of sample}

\item{raw_count_matrix_address}{address for raw count matrix}

\item{tissue_position_address}{address for tissue position matrix}

\item{background_image_address}{address for background address}

\item{giotto_python_path}{the path to a python which the Giotto can use}

\item{giotto_results_folder}{the folder for Giotto save plots}

\item{filtered_count_matrix}{address for filtered count matrix}
}
\description{
Initialize class Hematoma
}
