% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-ICH_Mouse.R
\name{initialize,ICH_Mouse-method}
\alias{initialize,ICH_Mouse-method}
\title{Initialize class ICH_Mouse}
\usage{
\S4method{initialize}{ICH_Mouse}(
  .Object,
  analysis_symbol,
  raw_count_matrix_address,
  filtered_count_matrix_address,
  tissue_position_address,
  background_image_address,
  giotto_python_path,
  giotto_results_folder
)
}
\description{
Initialize class ICH_Mouse
}
