% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-Save_Load_ICHMouse.R
\name{load_graph}
\alias{load_graph}
\title{load graph}
\usage{
load_graph(graph_name, data_symbol = "ich_mouse", loading_path = NULL)
}
\arguments{
\item{graph_name}{the name of the graph}

\item{data_symbol}{the data symbol}

\item{loading_path}{the path loading data}
}
\description{
load graph
}
