% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-Save_Load_ICHMouse.R
\name{Load_ICHMouse}
\alias{Load_ICHMouse}
\title{load ICHMouse data}
\usage{
Load_ICHMouse(class_symbol, data_symbol = "ich_mouse", loading_path = NULL)
}
\arguments{
\item{class_symbol}{the class type}

\item{data_symbol}{the data symbol}

\item{loading_path}{the path saving data}
}
\description{
load ICHMouse data
}
