% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-Plotting_Dataset.R
\docType{class}
\name{Plotting_Dataset-class}
\alias{Plotting_Dataset-class}
\title{the class for Plotting}
\description{
the class for Plotting
}
\section{Slots}{

\describe{
\item{\code{seurat_metadata_dataset}}{the Seurat Object metadata dataset for plotting}

\item{\code{differential_expression_genes_dataset}}{the differential expression genes dataset}

\item{\code{plots}}{the plots created from dataset}
}}

