% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-Save_ICHMouse.R
\name{Save_ICHMouse}
\alias{Save_ICHMouse}
\title{save ICHMouse data}
\usage{
Save_ICHMouse(Class_ICHMouse, saving_path = getwd())
}
\arguments{
\item{Class_ICHMouse}{the class of ICHMouse}

\item{saving_path}{the path for saving}
}
\description{
save ICHMouse data
}
